PART FINAL - Mobile registration of a new user
add a negative test case for invalid passwords, bug attached
add @BeforeClass, @afterclass, deletion of created users enhanced

+++ src/test/java/com/photopixels/base/IApiBaseTest.java
@@ -25,6 +25,7 @@ import java.security.NoSuchAlgorithmException;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
+import java.util.NoSuchElementException;
 
 import static com.photopixels.constants.Constants.SAMPLE_VIDEO_FILE;
 import static com.photopixels.constants.Constants.TRAINING_FILE;
@@ -85,7 +86,10 @@ public interface IApiBaseTest extends IBaseTest {
         GetUserResponseDto[] getUserResponseDtos = getUsersSteps.getUsers();
 
         return Arrays.stream(getUserResponseDtos)
-                .filter(user -> user.getEmail().equals(username)).findFirst().get().getId();
+                .filter(user -> user.getEmail().equals(username))
+                .map(GetUserResponseDto::getId)
+                .findFirst()
+                .orElse("");
     }
 
     default void deleteRegisteredUsers(Map<String, String> registeredUsers) {
@@ -105,6 +109,10 @@ public interface IApiBaseTest extends IBaseTest {
 
             for (String username : registeredUsers) {
                 String userId = getUserId(username);
+
+                if (userId.isEmpty()) {
+                    continue;
+                }
 
                 DeleteUserAdminSteps deleteUserAdminSteps = new DeleteUserAdminSteps(token);
                 deleteUserAdminSteps.deleteUserAdmin(userId);
